requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
bcrypt==4.0.1
aiofiles==23.2.1
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2
gunicorn==21.2.0
requests==2.31.0
